Feature: Add skeleton to the users' list
